refactor: Change Category ID to Long and add OrderRepository
This commit introduces several improvements and fixes across the application:

- The primary key for the Category entity has been changed from Integer to Long to prevent potential overflow and align with best practices. This change is propagated to the `ManagerCategoryController` and `ManagerProductController`.

- The `mssql-jdbc` dependency version is now explicitly defined in `pom.xml` to ensure build consistency and prevent potential conflicts.

- A default status of "Active" is now set for new User entities, ensuring data integrity upon creation.

- A new `OrderRepository` has been implemented with methods to find orders by user, enabling future features related to order history.

+++ src/main/java/com/flowerShop1/service/category/CategoryService.java
@@ -12,7 +12,7 @@ public interface CategoryService {
     List<Category> getAllCategory();
     List<CategoryDTO> getAllCategoriesWithProductCount();
     Page<Category> getAllCategories(String keyword, Pageable pageable);
-    Optional<Category> getById(Integer id);
+    Optional<Category> getById(Long id);
     Category save(Category category);
-    void deleteById(Integer id);
+    void deleteById(Long id);
 }
